more function renaming and moving around
