+++ package.json
@@ -54,14 +54,12 @@
     "istanbul": "^0.4.5",
     "markdownlint-cli": "^0.3.1",
     "minimist": "^1.2.0",
-    "mock-aws": "^1.2.3",
     "npm-run-all": "^4.0.2",
     "supertest": "^3.0.0",
     "unionized": "^4.11.0"
   },
   "dependencies": {
     "async": "^2.2.0",
-    "aws-s3-zipper": "^1.0.1",
     "aws-sdk": "^2.32.0",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.17.1",
@@ -81,6 +79,7 @@
     "matchstick": "^1.2.1",
     "mkdirp": "^0.5.1",
     "mocha": "^5.1.1",
+    "mock-aws-s3": "^2.6.0",
     "morgan": "^1.8.1",
     "moxai": "^1.1.1",
     "multer": "^1.3.0",
